requirement.txt
streamlit
pdfplumber
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
faiss-cpu
python-dotenv